Cdstp 120 separate old functionality into legacy folder with same structure (#107)
* CDSTP-120 Separated pv_to_puml into v1 and main

* CDSTP-120 Moved all legacy loop detection functionality to separate folder

* CDSTP-120 Changed naming convention for folder

* CDSTP-120 Fix linting and mypy errors

* CDSTP-120 Separated more legacy loop detection functionality

* CDSTP-120 updated test names of legacy

* CDSTP-120 Remove unused file
