+++ tests/tel2puml/legacy_loop_detection/puml_graph/conftest.py
@@ -5,10 +5,9 @@ import pytest
 from tel2puml.puml_graph.graph import (
     PUMLGraph,
     PUMLEventNode,
-    PUMLOperatorNode,
 )
 from tel2puml.tel2puml_types import PUMLOperator, PUMLEvent
-from tel2puml.loop_detection.loop_types import DUMMY_BREAK_EVENT_TYPE
+from tel2puml.legacy_loop_detection.detect_loops import Loop
 
 
 @pytest.fixture
@@ -96,53 +95,134 @@ def puml_graph() -> tuple[PUMLGraph, dict[tuple[str, int], PUMLEventNode]]:
 
 
 @pytest.fixture
-def graph_with_dummy_break_event(
-) -> tuple[PUMLGraph, dict[str, PUMLEventNode | PUMLOperatorNode]]:
-    """Return a PUMLGraph instance with a dummy break event.
+def loop_1() -> Loop:
+    """Return a Loop instance.
+
+    :return: Loop instance
+    :rtype: :class:`Loop`
+    """
+    loop = Loop(nodes=["q1", "q2", "q3", "q4"])
+    for end in ["q3", "q4"]:
+        for start in ["q1", "q2"]:
+            loop.add_edge_to_remove((end, start))
+    return loop
+
+
+@pytest.fixture
+def subloop() -> Loop:
+    """Return a Loop instance that is a subloop.
+
+    :return: Loop instance
+    :rtype: :class:`Loop`
+    """
+    loop = Loop(nodes=["q6"])
+    loop.add_edge_to_remove(("q6", "q6"))
+    return loop
+
+
+@pytest.fixture
+def loop_2(subloop: Loop) -> Loop:
+    """Return a Loop instance with a subloop.
+
+    :param subloop: The subloop to add to the loop.
+    :type subloop: :class:`Loop`
+    :return: Loop instance
+    :rtype: :class:`Loop`
+    """
+    loop = Loop(nodes=["q5", "q6", "q7"])
+    loop.add_edge_to_remove(("q7", "q5"))
+    loop.sub_loops.append(subloop)
+    return loop
+
+
+@pytest.fixture
+def expected_puml_graph_post_loop_insertion() -> PUMLGraph:
+    """Return a PUMLGraph instance after loop insertion.
 
     The graph is as follows:
     ```
-    START -> A -> XOR_START_1 -> B -> XOR_END_1
-    START -> A -> XOR_START_1 -> XOR_START_2 -> C -> XOR_END_2 -> XOR_END_1
-    START -> A -> XOR_START_1 -> XOR_START_2 -> DUMMY_BREAK_EVENT -> XOR_END_2\
-    -> XOR_END_1
+    A -> AND -> LOOP1 -> AND -> F
+    A -> AND -> LOOP2 -> AND -> F
+    A -> AND -> LOOP3 -> AND -> F
     ```
     :return: PUMLGraph instance
     :rtype: :class:`PUMLGraph`
     """
     graph = PUMLGraph()
-    START = graph.create_event_node(
-        "START", parent_graph_node="START"
-    )
     A = graph.create_event_node(
-        "A", parent_graph_node="A"
+        "A", parent_graph_node="q0"
     )
-    XOR_START_1, XOR_END_1 = graph.create_operator_node_pair(PUMLOperator.XOR)
-    B = graph.create_event_node("B", parent_graph_node="B")
-    XOR_START_2, XOR_END_2 = graph.create_operator_node_pair(PUMLOperator.XOR)
-    C = graph.create_event_node("C", parent_graph_node="C")
-    DUMMY_BREAK_EVENT = graph.create_event_node(
-        DUMMY_BREAK_EVENT_TYPE, event_types=PUMLEvent.BREAK
+    F = graph.create_event_node(
+        "F", parent_graph_node="q8"
     )
-    graph.add_edge(START, A)
-    graph.add_edge(A, XOR_START_1)
-    graph.add_edge(XOR_START_1, B)
-    graph.add_edge(XOR_START_1, XOR_START_2)
-    graph.add_edge(B, XOR_END_1)
-    graph.add_edge(XOR_START_2, C)
-    graph.add_edge(XOR_START_2, DUMMY_BREAK_EVENT)
-    graph.add_edge(C, XOR_END_2)
-    graph.add_edge(XOR_END_1, XOR_END_2)
-    graph.add_edge(DUMMY_BREAK_EVENT, XOR_END_2)
-    nodes: dict[str, PUMLEventNode | PUMLOperatorNode] = {
-        "START": START,
-        "A": A,
-        "B": B,
-        "C": C,
-        DUMMY_BREAK_EVENT_TYPE: DUMMY_BREAK_EVENT,
-        "XOR_START_1": XOR_START_1,
-        "XOR_END_1": XOR_END_1,
-        "XOR_START_2": XOR_START_2,
-        "XOR_END_2": XOR_END_2
-    }
-    return graph, nodes
+    LOOP1 = graph.create_event_node(
+        "LOOP", event_types=PUMLEvent.LOOP
+    )
+    LOOP2 = graph.create_event_node(
+        "LOOP", event_types=PUMLEvent.LOOP
+    )
+    LOOP3 = graph.create_event_node(
+        "LOOP", event_types=PUMLEvent.LOOP
+    )
+    and_start, and_end = graph.create_operator_node_pair(PUMLOperator.AND)
+    graph.add_edge(A, and_start)
+    graph.add_edge(and_end, F)
+    graph.add_edge(and_start, LOOP1)
+    graph.add_edge(and_start, LOOP2)
+    graph.add_edge(and_start, LOOP3)
+    graph.add_edge(LOOP1, and_end)
+    graph.add_edge(LOOP2, and_end)
+    graph.add_edge(LOOP3, and_end)
+    return graph
+
+
+@pytest.fixture
+def expected_loop_1_graph() -> PUMLGraph:
+    """Return a PUMLGraph instance for loop 1 that can be checked against the
+    actual graph obtained in the test.
+
+    The graph is as follows:
+    ```
+    XOR -> B -> D -> XOR
+    XOR -> C -> E -> XOR
+    ```
+    :return: PUMLGraph instance
+    :rtype: :class:`PUMLGraph`
+    """
+    graph = PUMLGraph()
+    xor_start, xor_end = graph.create_operator_node_pair(PUMLOperator.XOR)
+    B = graph.create_event_node("B", parent_graph_node="q1")
+    C = graph.create_event_node("C", parent_graph_node="q2")
+    D = graph.create_event_node("D", parent_graph_node="q3")
+    E = graph.create_event_node("E", parent_graph_node="q4")
+    graph.add_edge(xor_start, B)
+    graph.add_edge(xor_start, C)
+    graph.add_edge(B, D)
+    graph.add_edge(C, E)
+    graph.add_edge(D, xor_end)
+    graph.add_edge(E, xor_end)
+    return graph
+
+
+@pytest.fixture
+def expected_loop_2_graph() -> PUMLGraph:
+    """Return a PUMLGraph instance for loop 2 that can be checked against the
+    actual graph obtained in the test.
+
+    The graph is as follows:
+    ```
+    G -> LOOP -> I
+    ```
+
+    :return: PUMLGraph instance
+    :rtype: :class:`PUMLGraph`
+    """
+    graph = PUMLGraph()
+    G = graph.create_event_node("G", parent_graph_node="q5")
+    LOOP = graph.create_event_node(
+        "LOOP", event_types=PUMLEvent.LOOP
+    )
+    INODE = graph.create_event_node("I", parent_graph_node="q7")
+    graph.add_edge(G, LOOP)
+    graph.add_edge(LOOP, INODE)
+    return graph
