+++ tests/tel2puml/legacy_loop_detection/test_detect_loops_legacy.py
@@ -7,10 +7,11 @@ from networkx import DiGraph
 from tel2puml.data_pipelines.data_ingestion import (
     update_all_connections_from_clustered_events,
 )
-from tel2puml.events import (
-    events_to_markov_graph, get_event_reference_from_events
+from tel2puml.events import events_to_markov_graph
+from tel2puml.legacy_loop_detection.events import (
+    get_event_reference_from_events,
 )
-from tel2puml.detect_loops import (
+from tel2puml.legacy_loop_detection.detect_loops import (
     Loop,
     detect_loops,
     add_loop_edges_to_remove,
