+++ tests/tel2puml/loop_detection/test_loop_event_methods.py
@@ -140,7 +140,7 @@ class TestCreateLoopEvent:
     def test_create_loop_event(self) -> None:
         """Tests the create loop event method."""
         loop, graph = self.loop_and_graph()
-        loop_event = create_loop_event(loop, graph, graph)
+        loop_event = create_loop_event(loop, graph)
         assert loop_event.in_event_sets == {EventSet(["A"])}
         assert loop_event.event_sets == {EventSet(["E"])}
         assert loop_event.event_type == LOOP_EVENT_TYPE
